requirements.txt
asgiref==3.8.1
attrs==24.2.0
autopep8==2.3.1
black==24.10.0
click==8.1.7
Django==4.2.16
django-extensions==3.2.3
django-silk==5.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.1
gprof2dot==2024.6.6
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
ruff==0.8.2
sqlparse==0.5.2
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
zipp==3.20.2
